# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Management API",
        "description": "API documentation for the PMS backend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local API"
        }
    ],
    "paths": {
...
        "/projects/{project}/sprints": {
            "get": {
                "tags": [
                    "Sprints"
                ],
                "summary": "List sprints for a project",
                "description": "Returns a list of all sprints in the specified project.",
                "operationId": "listSprints",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sprints retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Create a new sprint",
                "description": "Creates a new sprint in the specified project.",
                "operationId": "createSprint",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "The ID of the project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The name of the sprint",
                                        "type": "string"
                                    },
                                    "goal": {
                                        "description": "The goal of the sprint",
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "description": "Start date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "description": "End date",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sprint created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/sprints/{sprint}": {
            "get": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Get sprint details",
                "description": "Returns details of a specific sprint.",
                "operationId": "getSprint",
                "parameters": [
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "The ID of the sprint",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sprint retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Update a sprint",
                "description": "Updates an existing sprint.",
                "operationId": "updateSprint",
                "parameters": [
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "The ID of the sprint",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The name of the sprint",
                                        "type": "string"
                                    },
                                    "goal": {
                                        "description": "The goal of the sprint",
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "description": "Start date",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "description": "End date",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sprint updated successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Delete a sprint",
                "description": "Deletes an existing sprint.",
                "operationId": "deleteSprint",
                "parameters": [
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "The ID of the sprint",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sprint deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/sprints/{sprint}/start": {
            "post": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Start a sprint",
                "description": "Starts the sprint.",
                "operationId": "startSprint",
                "parameters": [
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "The ID of the sprint",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sprint started successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/sprints/{sprint}/complete": {
            "post": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Complete a sprint",
                "description": "Completes the sprint.",
                "operationId": "completeSprint",
                "parameters": [
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "The ID of the sprint",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "carry_over_sprint_id": {
                                        "description": "ID of the sprint to carry over issues to",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sprint completed successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/sprints/{sprint}/issues/{issue}": {
            "post": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Add issue to sprint",
                "description": "Adds an issue to the sprint.",
                "operationId": "addIssueToSprint",
                "parameters": [
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "The ID of the sprint",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "issue",
                        "in": "path",
                        "description": "The ID of the issue",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Issue added to sprint"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sprints"
                ],
                "summary": "Remove issue from sprint",
                "description": "Removes an issue from the sprint.",
                "operationId": "removeIssueFromSprint",
                "parameters": [
                    {
                        "name": "sprint",
                        "in": "path",
                        "description": "The ID of the sprint",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "issue",
                        "in": "path",
                        "description": "The ID of the issue",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Issue removed from sprint"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}